+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/service/impl/ImageService.java
@@ -13,13 +13,13 @@ public class ImageService implements ImageServiceAware {
     private final ImageRepository imageRepository;
 
     @Override
-    public int saveImage(Image image) {
+    public Long saveImage(Image image) {
         Image savedImage = imageRepository.save(image);
         return savedImage.getId();
     }
 
     @Override
-    public Image findById(int id) throws NotFoundException {
+    public Image findById(Long id) throws NotFoundException {
         return imageRepository.findById(id).orElseThrow(
                 () -> new NotFoundException(String.format("Cannot find image by id = %d", id)));
     }
